backend/src/api/dependencies/dependencies.py
from datetime import datetime

from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from jose import jwt
from pydantic import ValidationError

from src.core.config import get_settings
from src.models.Patient import Patient
from src.models.User import User
from src.schemas.TokenSchema import TokenPayload
from src.services.PatientService import PatientService
from src.services.UserService import UserService

settings = get_settings()


reuseable_oauth = OAuth2PasswordBearer(
    tokenUrl=f"{settings.AP1_V1_STR}/users/login", scheme_name="JWT"
)


async def get_current_user(token: str = Depends(reuseable_oauth)) -> User:
    try:
        payload = jwt.decode(
            token, settings.JWT_SECRET_KEY, algorithms=[settings.ALGORITHM]
        )
        token_data = TokenPayload(**payload)

        if datetime.fromtimestamp(token_data.exp) < datetime.now():
            raise HTTPException(
                status_code=status.HTTP_401_UNAUTHORIZED,
                detail="Token expired",
                headers={"WWW-Authenticate": "Bearer"},
            )
    except (jwt.JWTError, ValidationError):
        raise HTTPException(
            status_code=status.HTTP_403_FORBIDDEN,
            detail="Could not validate credentials",
            headers={"WWW-Authenticate": "Bearer"},
        )

    user = await UserService.get_user_by_id(token_data.sub)
    if not user:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail="Could not find user",
        )

    return user


async def get_current_patient(user: User = Depends(get_current_user)) -> Patient:
    patient = await PatientService.get_patient_profile(user)
    return patient
