backend/requirements.txt
anyio==3.6.2
attrs==22.1.0
bcrypt==4.0.1
beanie==1.15.4
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
exceptiongroup==1.0.4
fastapi==0.86.0
fastapi-utils==0.2.1
greenlet==2.0.1
h11==0.14.0
idna==3.4
iniconfig==1.1.1
mccabe==0.7.0
motor==3.1.1
multidict==6.0.3
mypy-extensions==0.4.3
packaging==21.3
passlib==1.7.4
pathspec==0.10.2
PIL-Tools==1.1.0
Pillow==9.3.0
platformdirs==2.5.4
pluggy==1.0.0
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.2
pydocstyle==6.1.1
pyflakes==3.0.1
pylama==8.4.1
pymongo==4.3.2
pyparsing==3.0.9
pytest==7.2.0
python-decouple==3.6
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
SQLAlchemy==1.4.44
starlette==0.20.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.19.0
yarl==1.8.2